clip-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.65.0
python-multipart>=0.0.6
